main/requirements.txt
PyQt5==5.15.9
pyserial==3.5
firebase-admin==6.2.0
requests==2.31.0
python-escpos==3.0a9
opencv-python-headless==4.8.0.74
qrcode==7.4.2
Pillow==10.0.0
